+++ include/ArduinoJson/Internals/JsonInteger.hpp
@@ -14,10 +14,13 @@ namespace Internals {
 
 #if ARDUINOJSON_USE_LONG_LONG
 typedef long long JsonInteger;
+typedef unsigned long long JsonUInt;
 #elif ARDUINOJSON_USE_INT64
 typedef __int64 JsonInteger;
+typedef unsigned _int64 JsonUInt;
 #else
 typedef long JsonInteger;
+typedef unsigned long JsonUInt;
 #endif
 }
 }