+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ArduinoJson: change log
 =======================
 
-HEAD
-----
+v5.3.0
+------
 
 * Added custom implementation of `ftoa` (issues #266, #267, #269 and #270)
 * Added `JsonVariant JsonBuffer::parse()` (issue #265)